Initial Import - Calculate show/season-average from rated episodes
This first draft handles a very rudimentary authentication and then
fetches all rated episodes of the authenticated user. Using HashMaps,
a season/show average is then calculated. And finally, all results
are then printed to the standard output (usually the commandline).

This is not yet beautiful or even nice, but it works now.