Add property-handling and skeleton-property file
To better handle access tokens and device codes and to make them more
persistent, I've decided to implement property-handling using a
property file.

In the given skeleton-property file the first two values (secret and id)
are already pre-filled (seems it has to be this way? but seems weird...)
and the other two will be generated on the first run of the program.

I may have to remove the first two values in the future, but for now
they stay in.