code cleanup + new functionality: fetch rating of shows/seasons
I've dusted a few things and split the whole big method of fetching the
episode-ratings and calculating/printing into smaller, different
methods. And I added fetching the rating of shows/season.

I also tried to add jar-support but this worked only partially because
of the setting-file I use - so this won't be in this commit (only in the
README).